Move generic code to common package and clean up
