+++ pkg/servicecfg/cinder.go
@@ -20,6 +20,7 @@ import (
 	"io/ioutil"
 	"strings"
 
+	"github.com/openstack-k8s-operators/os-diff/pkg/godiff"
 	"gopkg.in/yaml.v3"
 )
 
@@ -89,5 +90,5 @@ func LoadCinderOpenShiftConfig(configPath string) string {
 		sb.WriteString(cinder.Spec.Cinder.Template.CustomServiceConfig)
 	}
 
-	return cleanIniSections(sb.String())
+	return godiff.CleanIniSections(sb.String())
 }
